voir les infos du user courant

+++ controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel");
 
 // @desc récupérer tous les contacts
 // @route GET /api/getContacts
-// @access public
+// @access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts)
 });
 
 // @desc récupérer un contact
 // @route GET /api/getContacts/:id
-// @access public
+// @access private
 const getContact = asyncHandler( async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
@@ -23,7 +23,7 @@ const getContact = asyncHandler( async (req, res) => {
 
 // @desc créer un nouveau contact
 // @route POST /api/getContacts/
-// @access public
+// @access private
 const createContact = asyncHandler( async (req, res) => {
     console.log(req.body);
     const { name, email, phone } = req.body;
@@ -36,6 +36,7 @@ const createContact = asyncHandler( async (req, res) => {
         name,
         email,
         phone,
+        user_id: req.user.id,
     });
     
     res.status(200).json(contact);
@@ -43,12 +44,17 @@ const createContact = asyncHandler( async (req, res) => {
 
 // @desc modifier un contact
 // @route PUT /api/getContacts/:id
-// @access public
+// @access private
 const updateContact = asyncHandler( async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("contact non trouvé");
+    }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("Vous n'avez pas la permission de modifier ce contact"); 
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -61,13 +67,19 @@ const updateContact = asyncHandler( async (req, res) => {
 
 // @desc supprimer un contact
 // @route DELETE /api/getContacts/:id
-// @access public
+// @access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("contact non trouvé");
     }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("Vous n'avez pas la permission de modifier ce contact"); 
+    }
+    
     await Contact.deleteOne();
     console.log("contact");
     res.status(200).json(contact);
